Updated login and register pages with some form items for submission. Committing for pull/merge purposes.

+++ client/src/components/Body/DashboardPage/Students/StudentTable.jsx
@@ -5,16 +5,20 @@ import '../../../../styles/Students.css';
 
 const StudentTable = ({ stud, onAdd }) => {
 
-    function handleDelete(num) {
-      let studNum = parseInt(num);
-      fetch(`http://localhost:8000/students/${studNum}`, 
-      { 
-          method: 'DELETE',
-          credentials: "same-origin",
-          headers: {"Content-Type": 'application/json'}
-      })
-      .then(response => response.json())
-      .then(onAdd());
+    
+
+    function handleDelete(num, name) {
+      if (confirm(`Are you sure you want to delete ${name}?`)) {
+        const studNum = parseInt(num);
+        fetch(`http://localhost:8000/students/${studNum}`, 
+        { 
+            method: 'DELETE',
+            credentials: "same-origin",
+            headers: {"Content-Type": 'application/json'}
+        })
+        .then(response => response.json())
+        .then(onAdd());      
+      }
     }
   
     return (
@@ -26,19 +30,20 @@ const StudentTable = ({ stud, onAdd }) => {
               <th>GitHub</th>
               <th>Graduation Date</th>
               <th>Cohort</th>
-              <th></th>
             </tr>
             {stud.map((val, key) => {
                 return (
+                  <>
                     <tr className='table-row' key={`Student${key}`} onClick={()=> document.getElementById(`Student${key}`).style.display = "flex"}>
                         <td>{val.stu_name}</td>
                         <td>{val.email}</td>
                         <td>{val.github}</td>
                         <td>{val.graduation}</td>
                         <td>{val.cohort_number}</td>
-                        <td><button className='delete-student' onClick={()=> handleDelete(val.id)}>Delete</button></td>
-                        <UpdateStudent keyID={`Student${key}`} stud={val} onAdd={onAdd} />
                     </tr>
+                    <button className='delete-student' onClick={()=> handleDelete(val.id, val.stu_name)}>Delete</button>
+                    <UpdateStudent keyID={`Student${key}`} stud={val} onAdd={onAdd} />
+                  </>
                 )
             })}
         </table>
